requirements.txt
numpy
ftfy
regex
hydra-core
wandb
lightning
tensorboard
clip @ git+https://github.com/openai/CLIP.git
tqdm
Pillow
